backend/requirements.txt
fastapi==0.135.2
uvicorn==0.42.0
python-dotenv==1.0.1
requests==2.31.0
pandas==3.0.2
numpy==2.1.0
scikit-learn==1.6.1
joblib==1.5.3